+++ changelog.md
@@ -2,7 +2,11 @@
 
 #### v2.7.5 (2026-06-xx)
 
+Thank you to our amazing contributors for helping out. Thanks!
+- Moosi_13 for reporting Priest and Paladin issues
+
 The following glowing buttons have been added temporarily (Mists of Pandaria)
+- Paladin's Flash of Light, while Selfless Healer is at 3 stacks
 - Priest's Shadow Word: Death
 - Priest's Power Word: Shield, during Divine Insight
 - Priest's Praryer of Mending, during Divine Insight